requirements.txt
streamlit
requests
pandas
beautifulsoup4
PyMuPDF
openai
python-dotenv
tqdm
typing_extensions

# Google Gemini для глубокого сканирования
google-generativeai

# Threads API dependencies
aiohttp
cryptography
colorama
